Make links work in the QtPdf backend
This is a substantial rewrite now that I have a better understanding
of both QPdfLinkModel and the original Poppler code.

Links are now placed in the correct areas on the page. In addition,
the Link.url property now returns a QString, as it does under the
Poppler backend, rather than a QUrl.

So far as I've been able to test, this brings the QtPdf backend's
link support up to par with the Poppler backend's.