Correct deformed PDF rendering on zoom > 125% (fixes #25)
Above this zoom level, the image is broken up into individual tiles
that are rendered separately. The deformation was caused by the
QtPdf backend handling the tile geometry incorrectly.

Whereas Poppler allows rendering only a specified area of a page,
QtPdf forces us to render the entire page. We can emulate Poppler's
selective rendering by cropping our own rendered image to the tile
boundaries; this preserves compatibility with the existing logic,
but is a very inefficient approach that largely defeats the purpose
of tile-based rendering.