Trigger oversampling based on the paint device's resolution
The definition of oversampleThreshold = 96 goes back to the very first
commit of poppler.py back when qpageview was still part of Frescobaldi.
No explanation is given, but I'm presuming it's because 96 dpi is the
traditional standard pixel density for PC screens.

In practice, painter.device().logicalDpi[XY]() seems to return 96 anyway
on both standard and high-DPI screens, but based on my understanding of
the scaling logic this is more semantically correct.