Fix local path normalization on Windows
Capitalizing the drive letter here fixes Frescobaldi opening a separate
copy of the document when using point-and-click. While the actual bug
seems to involve code elsewhere assuming case sensitivity, working
around it this way is easier and has much less chance of breaking things
on platforms that do use case-sensitive paths.

Since I was reworking this section anyway, I also rewrote the logic to
restore a missing colon after the drive letter to make it easier to
understand.