Fix two potential issues with qpageview.loadPdf()
These have not yet been problems in practice, but they would be
if someone was to create a python-poppler-qt6 module.

The first fix is to force the use of QtPdf when the filename
argument is a QPdfDocument (a QtPdf-specific type).

The second, smaller fix is to move the check for popplerqt6 into
loadPdf() since that is the only function that needs it.