Remove remaining pieces that depend on Poppler
This most obviously includes the Poppler backend, but also the EPS
exporter.

As discussed under #32, there is currently no python-poppler-qt6, and
little interest in creating one now that QtPdf satisfies our needs. It
would therefore make sense to drop Poppler support rather than continue
to include a feature we do not need using a binding we do not control.

This PR is mutually exclusive with #33 and intended to follow
frescobaldi/frescobaldi#1832, which removes the last Poppler-dependent
features from Frescobaldi.